requirements.txt
numpy
pillow
torch
git+https://github.com/openai/CLIP.git
